feature: skip optional private endpoints without failing
This change is intended for optional dependencies that leave their
configuration parameters in the "private endpoints" section. When
applications depend on us, and they try to deploy everything to an
environment, they might end up without transitive dependencies, if these
are specified in the "optional dependencies" section of our ClowdApp.
This can cause the configuration reader to fail because the
configurations of those transitive dependencies will not be present,
which in turn stops the engine from booting.

We recently had an issue where the provisioning team was trying to
deploy their application, which depended on Notifications, in the
ephemeral environment. Since the export service was defined as an
optional dependency, it was not being deployed along with Notifications.
This caused the configuration reader to fail, because of the missing
configuration. Therefore, we need a way to allow skipping certain
configuration values for these optional transitive dependencies.

RHCLOUD-27017